refactor: extract JST date helpers and config-check helpers
- Create src/lib/jst.ts with getJSTNow(), getJSTToday(), getJSTTomorrow()
- Replace inline JST calculation in handler.ts and local functions in scheduler.ts
- Add requireProjectNumber() and requireRepo() helpers to handler.ts
  to consolidate 6 repetitive config validation blocks
